koiben-api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
motor>=3.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
certifi>=2023.7.22
dnspython>=2.3.0
